Fixes image refresh after loving a post or editorial.

+++ Sources/Controllers/Stream/StreamDataSource.swift
@@ -452,9 +452,9 @@ class StreamDataSource: NSObject, UICollectionViewDataSource {
             }
 
         case .delete:
+            var removedAvatar = false
             if let love = jsonable as? Love {
                 if let post = love.post, let user = love.user {
-                    var removed = false
                     for item in visibleCellItems {
                         if let userAvatars = item.jsonable as? UserAvatarCellModel,
                             userAvatars.belongsTo(post: post, type: .lovers)
@@ -462,22 +462,22 @@ class StreamDataSource: NSObject, UICollectionViewDataSource {
                             userAvatars.remove(user: user)
                             if userAvatars.users.count == 0 {
                                 replacePlaceholder(type: .postLovers, items: [])
-                                removed = true
+                                removedAvatar = true
                             }
                             break
                         }
                     }
 
-                    if removed && !hasCellItems(for: .postReposters) {
+                    if removedAvatar && !hasCellItems(for: .postReposters) {
                         replacePlaceholder(type: .postSocialPadding, items: [])
                     }
-
-                    collectionView.reloadData()
                 }
             }
 
-            removeItemsFor(jsonable: jsonable, change: change)
-            collectionView.reloadData() // deleteItemsAtIndexPaths(indexPaths)
+            let removedPaths = removeItemsFor(jsonable: jsonable, change: change)
+            if removedAvatar || removedPaths.count > 0 {
+                collectionView.reloadData() // deleteItemsAtIndexPaths(indexPaths)
+            }
         case .replaced:
             let (oldIndexPaths, _) = elementsFor(jsonable: jsonable, change: change)
             if let post = jsonable as? Post, let firstIndexPath = oldIndexPaths.first {
